Move tooling to numba.misc.inspection
